+++ backend/src/main/java/coffeeshout/global/nickname/WordPicker.java
@@ -1,4 +1,4 @@
-package coffeeshout.room.domain.service;
+package coffeeshout.global.nickname;
 
 import java.util.List;
