+++ backend/src/main/java/coffeeshout/room/config/RoomConfig.java
@@ -1,6 +1,6 @@
 package coffeeshout.room.config;
 
-import coffeeshout.room.domain.service.WordPicker;
+import coffeeshout.global.nickname.WordPicker;
 import com.google.genai.Client;
 import com.vane.badwordfiltering.BadWordFiltering;
 import java.util.concurrent.ThreadLocalRandom;
