+++ backend/src/test/java/coffeeshout/room/domain/service/PlayerNameGeneratorTest.java
@@ -3,6 +3,7 @@ package coffeeshout.room.domain.service;
 import static coffeeshout.global.ExceptionAssertions.assertCoffeeShoutException;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import coffeeshout.global.nickname.WordPicker;
 import coffeeshout.room.domain.RoomErrorCode;
 import java.util.List;
 import java.util.Set;
